BUG/MAJOR: stats-file: fix crash on non-x86 platform caused by unaligned cast
Since commit d655ed5f14 ("BUG/MAJOR: stats-file: ensure
shm_stats_file_object struct mapping consistency (2nd attempt)"), the
last_state_change field in the counters is a uint (to match how it's
reported). However, it happens that there are explicit casts in function
me_generate_field() to retrieve the value, and which cause crashes on
aarch64 and likely other non-x86 64-bit platforms due to atomically
reading an unaligned 64-bit value, and may even randomly crash other
64-bit platforms when reading past the end of the structure.

The fix for now adapts the cast to match the one used by the accessed
type (i.e. unsigned int), but the approach must change, as there's
nothing there which allows to figure whether or not the type is correct
by just reading the code. At minima a typeof() on a named field is
needed, but this requires more invasive changes, hence this temporary
fix.

No backport is needed, as stats-file is only in 3.3.

+++ src/stats-proxy.c
@@ -376,12 +376,12 @@ static struct field me_generate_field(const struct stat_col *col,
 			value = mkf_u32(FN_RATE, read_freq_ctr(counter));
 	}
 	else if (fn == FN_AGE) {
-		unsigned long age;
+		unsigned int age;
 
 		if (col->flags & STAT_COL_FL_SHARED)
-			age = COUNTERS_SHARED_LAST_OFFSET(((char **)counter), unsigned long, offset);
+			age = COUNTERS_SHARED_LAST_OFFSET(((char **)counter), unsigned int, offset);
 		else
-			age = *(unsigned long *)counter;
+			age = *(unsigned int *)counter;
 
 		if (age)
 			age = ns_to_sec(now_ns) - age;
